configuracion del redux correcta
