+++ redux/store.js
@@ -1,13 +1,13 @@
 import {configureStore} from "@reduxjs/toolkit";
-import thunk from 'redux-thunk';
-import imagenesSlice from "./imagenes/imagenesSlice.js";
+
+//Slices
+import imagenesSlice from "./slices/imagenes/imagenesSlice.js";
 
 
-const store = configureStore({
+export const store = configureStore({
     reducer:{
         imagenes: imagenesSlice,
     },
-    middleware: [thunk]
 })
 
-export default store;
+
